proxy: Fix persistent session with oVirt 3.6
commit d3683c2 "proxy: Only set 'Prefer: persistent-auth' with
jsession-id"  was working as expected when using the
OvirtProxy::jsession-id and OvirtProxy::sso-token properties.

However, when not using these properties and expecting the jsessionid to
be set as a cookie during a regular HTTP session, this was not working
properly as oVirt is not sending a jsessionid cookie when Prefer:
persistent-auth is not set.

Consequently, we need to set it upon OvirtProxy creation in order not to
break this usecase. We don't add the header in RestProxyCall as was done
before as we want to be able to disable this addition after
OvirtProxy::sso-token has been set.